Chatbot_SmartVanBanQangNgai/requirements_local.txt
streamlit
langchain
langchain-community
langchain-huggingface
transformers
sentence-transformers
torch
docx2txt
pypdf
python-dotenv
